+++ EnjoyLife/src/com/blog/model/BlogDAO.java
@@ -18,8 +18,7 @@ public interface BlogDAO {
 	
 	public abstract boolean updateViewAmount(String postNo);
 	
-	public abstract boolean updateAvgScore(int score, String postNo);
-	
+	public abstract boolean updateAvgScore(double score, String postNo);
 	public List<BlogVO> selectAll();
 	public List<BlogVO> selectByFlagReport(int page);
 	public boolean reportChange(String flag, String postNo);
